+++ internal/parser/strict.go
@@ -3,7 +3,6 @@ package parser
 import (
 	"errors"
 	"fmt"
-	"strconv"
 	"strings"
 
 	"go.yaml.in/yaml/v3"
@@ -74,13 +73,13 @@ func (p Parser) parseGroups(doc *yaml.Node, offsetLine, offsetColumn int, conten
 			}
 			for _, group := range unpackNodes(entry.val) {
 				g := p.parseGroup(group, offsetLine, offsetColumn, contentLines)
-				if _, ok := names[g.Name]; ok {
+				if _, ok := names[g.Name.Value]; ok {
 					return nil, ParseError{
 						Line: group.Line,
 						Err:  errors.New("duplicated group name"),
 					}
 				}
-				names[g.Name] = struct{}{}
+				names[g.Name.Value] = struct{}{}
 				groups = append(groups, g)
 			}
 		}
@@ -116,7 +115,10 @@ func (p Parser) parseGroup(node *yaml.Node, offsetLine, offsetColumn int, conten
 				}
 				return group
 			}
-			group.Name = entry.val.Value
+			group.Name = YamlNode{
+				Value: entry.val.Value,
+				Pos:   diags.NewPositionRange(contentLines, entry.val, 1),
+			}
 		case "interval":
 			if entry.val.Kind != yaml.ScalarNode || entry.val.ShortTag() != strTag {
 				group.Error = ParseError{
@@ -126,10 +128,10 @@ func (p Parser) parseGroup(node *yaml.Node, offsetLine, offsetColumn int, conten
 				return group
 			}
 			group.Interval = newYamlDuration(entry.val, offsetLine, offsetColumn, contentLines, 1)
-			if group.Interval.Error != nil {
+			if group.Interval.ParseError != nil {
 				group.Error = ParseError{
 					Line: entry.key.Line,
-					Err:  fmt.Errorf("invalid %s value: %w", entry.key.Value, group.Interval.Error),
+					Err:  fmt.Errorf("invalid %s value: %w", entry.key.Value, group.Interval.ParseError),
 				}
 				return group
 			}
@@ -142,10 +144,10 @@ func (p Parser) parseGroup(node *yaml.Node, offsetLine, offsetColumn int, conten
 				return group
 			}
 			group.QueryOffset = newYamlDuration(entry.val, offsetLine, offsetColumn, contentLines, 1)
-			if group.QueryOffset.Error != nil {
+			if group.QueryOffset.ParseError != nil {
 				group.Error = ParseError{
 					Line: entry.key.Line,
-					Err:  fmt.Errorf("invalid %s value: %w", entry.key.Value, group.QueryOffset.Error),
+					Err:  fmt.Errorf("invalid %s value: %w", entry.key.Value, group.QueryOffset.ParseError),
 				}
 				return group
 			}
@@ -157,7 +159,7 @@ func (p Parser) parseGroup(node *yaml.Node, offsetLine, offsetColumn int, conten
 				}
 				return group
 			}
-			group.Limit, _ = strconv.Atoi(nodeValue(entry.val))
+			group.Limit = newYamlInt(entry.val, offsetLine, offsetColumn, contentLines, 1)
 		case "labels":
 			if entry.val.ShortTag() != mapTag {
 				group.Error = ParseError{